Copy source verbatim from exchange 'statistics' module
As part of extracting the statistics module into its own repository,
this commit copies verbatim the contents of the bisq-network/exchange
repository's 'statistics' module here into the root of the new
bisq-statsnode repository.

These sources cannot be built in their current form, due to issues with
the pom that will be taken care of in subsequent commits.