Organize imports
This change adds a .idea/codeStyles/Project.xml file that orders all
imports from higher-level layers to lower-level layers, with blank lines
between different major packages.

The purpose of this approach is to allow the reader to scan the imports
in a file efficiently to understand what kind of work the class does,
what layers it touches, and so forth. Keeping imports clean and ordered
by layer can be a powerful tool in keeping a codebase well organized and
tangle-free over time.

bisq.* packages are always sorted at the top.